rename teaspoons to terralab
