+++ .github/workflows/build-and-test.yml
@@ -55,7 +55,7 @@ jobs:
 
       - name: Run tests with coverage, write coverage report to coverage.xml
         run: |
-          poetry run pytest --cov-report xml --cov=teaspoons
+          poetry run pytest --cov-report xml --cov=terralab
 
       # The SonarQube scan is done here, so it can upload the coverage report generated by the tests.
       - name: SonarCloud Scan
@@ -77,7 +77,7 @@ jobs:
         with:
           status: failure
           channel: "#terra-tsps-alerts"
-          username: "Teaspoons CLI push to main branch"
+          username: "Terralab CLI push to main branch"
           author_name: "build-and-test-python-cli"
           icon_emoji: ":triangular_ruler:"
           fields: job, commit
