+++ .github/workflows/publish-cli.yml
@@ -46,6 +46,11 @@ jobs:
       - name: Install Poetry
         uses: snok/install-poetry@v1
 
-      - name: Build and publish the CLI
+      - name: Build the CLI using poetry
         run: |
-          poetry publish --build
+          poetry build
+      
+      - name: Publish the CLI to PyPi
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          packages-dir: dist/
