+++ .teaspoons-cli-config
@@ -1,4 +1,4 @@
-# Teaspoons API URL
+# Teaspoons service API URL
 TEASPOONS_API_URL=https://tsps.dsde-dev.broadinstitute.org
 
 # Port to use for local server (for auth)
@@ -8,6 +8,6 @@ SERVER_PORT=10444
 OAUTH_OPENID_CONFIGURATION_URI=https://terradevb2c.b2clogin.com/terradevb2c.onmicrosoft.com/b2c_1a_signup_signin_dev/v2.0/.well-known/openid-configuration
 OAUTH_CLIENT_ID=bbd07d43-01cb-4b69-8fd0-5746d9a5c9fe
 
-# Teaspoons storage (absolute path or relative to user's home directory)
-LOCAL_STORAGE_PATH=.teaspoons
+# Terralab storage (absolute path or relative to user's home directory)
+LOCAL_STORAGE_PATH=.terralab
