+++ README.md
@@ -6,8 +6,8 @@
 ## Python CLI structure
 The CLI code is structured as follows:
 ```
-teaspoons-cli
-├── teaspoons
+terra-scientific-pipelines-service-cli
+├── terralab
 │   └── commands
 │   │   └── __init__.py
 │   │   └── auth_commands.py
@@ -29,23 +29,29 @@ teaspoons-cli
 │   └── logic
 │   │   └── test_auth_logic.py
 │   │   └── test_pipelines_logic.py
+│   └── __init__.py
+│   └── auth_helper.py
+│   └── cli.py
+│   └── client.py
+│   └── config.py
+│   └── terralab
 ├── .gitignore
-├── .teaspoons-cli-config
+├── .terralab-cli-config
 ├── poetry.lock
 ├── pyproject.toml
 ├── README.md
 ```
 
-In the `teaspoons` directory, we have the following files and subdirectories:
-- `auth_helper.py` contains the code for authenticating with the Teaspoons service (Terra, via b2c).
+In the `terralab` directory, we have the following files and subdirectories:
+- `auth_helper.py` contains the code for authenticating with the Terra Scientific Pipelines Service (Terra, via b2c).
 - `cli.py` assembles the CLI sub-modules that are defined in `commands/`.
-- `client.py` contains the code for wrapping API calls to the Teaspoons service.
+- `client.py` contains the code for wrapping API calls to the Terra Scientific Pipelines Service.
 - `config.py` contains the code for managing the CLI configuration via environment variables.
-- `teaspoons` is the entrypoint for the CLI. It contains the main function that is called when the CLI is run.
+- `terralab` is the entrypoint for the CLI. It contains the main function that is called when the CLI is run.
 - `utils.py` contains utility functions that are used across the CLI.
 - The `commands` directory contains the CLI sub-modules. This is effectively the controller layer for the CLI.
-- The `logic` directory contains the business logic for the CLI, including calling Teaspoons APIs via the thin `teaspoons_client`, 
-which is autogenerated and published by the [Teaspoons service repository](https://github.com/DataBiosphere/terra-scientific-pipelines-service).
+- The `logic` directory contains the business logic for the CLI, including calling Terra Scientific Pipelines Service APIs via the thin `teaspoons_client`, 
+which is autogenerated and published by the [Terra Scientific Pipelines Service repository](https://github.com/DataBiosphere/terra-scientific-pipelines-service).
 
 In the `tests` directory, we have test files that can be run with pytest.
 
@@ -54,17 +60,17 @@ For now, the CLI requires poetry to be installed to run. See the [Development](#
 
 To run the CLI, navigate to the root (terra-scientific-pipelines-service-cli) directory and run the following command:
 ```bash
-./teaspoons COMMAND [ARGS]
+./terralab COMMAND [ARGS]
 ```
 
-For example, to authenticate with the Teaspoons service, run the following command:
+For example, to authenticate with Terralab (via Terra b2c), run the following command:
 ```bash
-./teaspoons auth login
+./terralab auth login
 ```
 
-To list the pipelines in the Teaspoons service, run the following command:
+To list the pipelines you can run using Terralab, run the following command:
 ```bash
-./teaspoons pipelines list
+./terralab pipelines list
 ```
 
 See WIP documentation for the CLI [here](https://docs.google.com/document/d/1ovbcHCzdyuC8RjFfkVJZiuDTQ_UAVrglSxSGaZwppoY/edit?tab=t.0#heading=h.jfsr3j3x0zjr).
@@ -90,7 +96,7 @@ poetry run pytest
 
 To run tests with a coverage report printed to the terminal:
 ```bash
-poetry run pytest --cov-report term --cov=teaspoons
+poetry run pytest --cov-report term --cov=terralab
 ```
 
 To run the formatter, execute the following command from the root project directory:
