+++ terralab/cli.py
@@ -3,9 +3,9 @@
 import click
 import logging
 
-from teaspoons import __version__, log
-from teaspoons.commands.auth_commands import auth
-from teaspoons.commands.pipelines_commands import pipelines
+from terralab import __version__, log
+from terralab.commands.auth_commands import auth
+from terralab.commands.pipelines_commands import pipelines
 
 
 # Context settings for commands, for overwriting some click defaults
@@ -14,7 +14,7 @@ CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 LOGGER = logging.getLogger(__name__)
 
 
-@click.group(name="teaspoons", context_settings=CONTEXT_SETTINGS)
+@click.group(name="terralab", context_settings=CONTEXT_SETTINGS)
 @click.version_option(__version__)
 @click.option(
     "--debug",
