+++ tests/.test.config
@@ -1,4 +1,4 @@
-# Teaspoons API URL
+# Teaspoons service API URL
 TEASPOONS_API_URL=not-real
 
 # Port to use for local server (for auth)
@@ -8,6 +8,6 @@ SERVER_PORT=12345
 OAUTH_OPENID_CONFIGURATION_URI=https://dontcare
 OAUTH_CLIENT_ID=whatever
 
-# Teaspoons storage (absolute path or relative to user's home directory)
+# Terralab storage (absolute path or relative to user's home directory)
 LOCAL_STORAGE_PATH=.cool
